PLANNER-625 Move optional example code into an example package (#1029)

+++ optaplanner-examples/src/main/java/org/optaplanner/examples/common/swingui/latitudelongitude/LatitudeLongitudeTranslator.java
@@ -84,7 +84,7 @@ public class LatitudeLongitudeTranslator {
     }
 
     public double translateYToLatitude(double y) {
-        return maximumLatitude - ((((double) y) - innerHeightMargin) * latitudeLength / innerHeight);
+        return maximumLatitude - ((y - innerHeightMargin) * latitudeLength / innerHeight);
     }
 
     public int getImageWidth() {
